+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - optimized MCP server initialization
 - optimized Aider connector start-up management
+- better detection of state for Quick Flow BMAD workflows
 
 ## [0.52.0]